# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sistema de Controle de Visitantes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visitors:
    get:
      operationId: listVisitors
      tags: [visitors]
      summary: List visitors with optional search
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: cpf
          schema:
            type: string
        - in: query
          name: phone
          schema:
            type: string
        - in: query
          name: company
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of visitors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorListResponse"
    post:
      operationId: createVisitor
      tags: [visitors]
      summary: Create visitor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
  /visitors/search:
    get:
      operationId: searchVisitors
      tags: [visitors]
      summary: Real-time search for autocomplete during visit registration
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matched visitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visitor"
  /visitors/{id}:
    get:
      operationId: getVisitor
      tags: [visitors]
      summary: Get visitor with full visit history
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Visitor with history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorWithHistory"
    patch:
      operationId: updateVisitor
      tags: [visitors]
      summary: Update visitor
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"

  # ── Visits ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    UserPermissions:
      type: object
      required:
        [
          viewDashboard,
          viewDashboardCharts,
          viewVisits,
          registerVisit,
          checkoutVisit,
          editVisit,
          cancelVisit,
          reprintLabel,
          viewVisitors,
          createVisitor,
          editVisitorName,
          editVisitorCpf,
          editVisitorBirthDate,
          editVisitorPhone,
          editVisitorCompany,
          editVisitorCity,
          accessServiceCenter,
          viewReports,
          manageSectors,
          manageSettings,
          viewAudit,
        ]
      properties:
        viewDashboard:
          type: boolean
        viewDashboardCharts:
          type: boolean
        viewVisits:
          type: boolean
        registerVisit:
          type: boolean
        checkoutVisit:
          type: boolean
        editVisit:
          type: boolean
        cancelVisit:
          type: boolean
        reprintLabel:
          type: boolean
        viewVisitors:
          type: boolean
        createVisitor:
          type: boolean
        editVisitorName:
          type: boolean
        editVisitorCpf:
          type: boolean
        editVisitorBirthDate:
          type: boolean
        editVisitorPhone:
          type: boolean
        editVisitorCompany:
          type: boolean
        editVisitorCity:
          type: boolean
        accessServiceCenter:
          type: boolean
        viewReports:
          type: boolean
        manageSectors:
          type: boolean
        manageSettings:
          type: boolean
        viewAudit:
          type: boolean
    User:
      type: object
      required:
        [
          id,
          name,
          login,
          role,
          status,
          mustChangePassword,
          permissions,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        login:
          type: string
        role:
          type: string
          enum: [admin, receptionist, attendant]
        sectorId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive]
        mustChangePassword:
          type: boolean
        permissions:
          $ref: "#/components/schemas/UserPermissions"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    Sector:
      type: object
      required:
        [
          id,
          name,
          abbreviation,
          secretariat,
          status,
          queueEnabled,
          usesDesks,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        secretariat:
          type: string
        status:
          type: string
          enum: [active, inactive]
        queueEnabled:
          type: boolean
        usesDesks:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Visitor:
      type: object
      required: [id, name, cpf, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        birthDate:
          description: Data de nascimento (AAAA-MM-DD); nula em cadastros antigos.
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    VisitorInput:
      type: object
      required: [name, cpf, birthDate]
      properties:
        name:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        phone:
          type: string
        company:
          type: string
        city:
          type: string
    VisitorUpdate:
      type: object
      properties:
        name:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        phone:
          type: string
        company:
          type: string
        city:
          type: string
    VisitorWithHistory:
      type: object
      required: [id, name, cpf, createdAt, visits]
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
        visits:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
    VisitorListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Visitor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Visit ─────────────────────────────────────────────────────────────
    Visit:
      type: object
      required:
        [
          id,
          visitorId,
          sectorId,
          status,
          entryDate,
          entryTime,
          entryUserId,
          createdAt,
        ]
      properties:
        id:
          type: integer
        visitorId:
          type: integer
        visitor:
          $ref: "#/components/schemas/Visitor"
        sectorId:
          type: integer
        sector:
          $ref: "#/components/schemas/Sector"
        responsible:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [waiting, ongoing, finished, cancelled]
        entryDate:
          type: string
          format: date
        entryTime:
          type: string
        entryUserId:
          type: integer
        entryUser:
          $ref: "#/components/schemas/User"
        exitDate:
          type: ["string", "null"]
          format: date
        exitTime:
          type: ["string", "null"]
        exitUserId:
          type: ["integer", "null"]
        exitUser:
          $ref: "#/components/schemas/User"
        cancelReason:
          type: ["string", "null"]
        priorityReason:
          description: Motivo de prioridade informado pela recepção.
          type: ["string", "null"]
        service:
          oneOf:
            - $ref: "#/components/schemas/VisitService"
            - type: "null"
        createdAt:
          type: string
          format: date-time
    VisitService:
      description: Situação da visita na fila da Central de Atendimento.
      type: object
      required: [status, queuedAt, priorityLevel]
      properties:
        status:
          type: string
          enum: [waiting, called, completed, cancelled]
        queuedAt:
          type: string
          format: date-time
        calledAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        priorityLevel:
          description: 0 = fila comum, 1 = prioritário, 2 = prioridade especial (80+).
          type: integer
        priorityReason:
          type: ["string", "null"]
        deskName:
          type: ["string", "null"]
        attendantName:
          type: ["string", "null"]
...