# packages/api-zod/src/generated/types/visitService.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Controle de Visitantes
 * OpenAPI spec version: 0.1.0
 */
import type { VisitServiceStatus } from './visitServiceStatus';

/**
 * Situação da visita na fila da Central de Atendimento.
 */
export interface VisitService {
  status: VisitServiceStatus;
  queuedAt: Date;
  /** @nullable */
  calledAt?: Date | null;
  /** @nullable */
  completedAt?: Date | null;
  /** 0 = fila comum, 1 = prioritário, 2 = prioridade especial (80+). */
  priorityLevel: number;
  /** @nullable */
  priorityReason?: string | null;
  /** @nullable */
  deskName?: string | null;
  /** @nullable */
  attendantName?: string | null;
}