# packages/api-zod/src/generated/types/visitorInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Controle de Visitantes
 * OpenAPI spec version: 0.1.0
 */

export interface VisitorInput {
  name: string;
  cpf: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  birthDate: string;
  phone?: string;
  company?: string;
  city?: string;
}