# packages/api-zod/src/generated/types/visitorWithHistory.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Controle de Visitantes
 * OpenAPI spec version: 0.1.0
 */
import type { Visit } from './visit';

export interface VisitorWithHistory {
  id: number;
  name: string;
  cpf: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  birthDate?: string | null;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
  visits: Visit[];
}